refactor(tui): separate emitted surface views
