+++ packages/daemon/src/tui/mirror/features/missions-activity/feature.tsx
@@ -1,6 +1,6 @@
 /* @jsxImportSource @opentui/solid */
 export { MissionsSurface, type MissionSurfaceProps } from "../../missions-surface.tsx";
-export { ActivitySurface, type ActivitySurfaceProps } from "../../activity-surface.tsx";
+export { ActivitySurface, type ActivitySurfaceProps } from "../../activity-surface-view.tsx";
 export { createMissionsActivityFeatureSession } from "./session.ts";
 export type {
   MissionsActivityFeatureHost,
