+++ packages/daemon/src/tui/mirror/files-surface-view.tsx
@@ -2,7 +2,7 @@
 import { For, Show } from "solid-js";
 import type { RGBA } from "@opentui/core";
 import type { SemanticThemeSnapshot } from "./theme.ts";
-import { ActionChip, EmptyState, InputShell, SelectableRow } from "./recipes.tsx";
+import { ActionChip, EmptyState, InputShell, SelectableRow } from "./recipe-components.tsx";
 import { recipePalette } from "./recipes.ts";
 import {
   filesHitTest,
