+++ packages/daemon/src/tui/mirror/home-files-surface-renderer.test.tsx
@@ -6,7 +6,7 @@ import { afterEach, describe, expect, it } from "bun:test";
 import { RGBA } from "@opentui/core";
 import { HomeSurface, homeActionAtProjection } from "./home-surface.tsx";
 import { projectHomeSurface } from "./home-surface.ts";
-import { FilesSurface } from "./files-surface.tsx";
+import { FilesSurface } from "./files-surface-view.tsx";
 import { filesHitTest, projectFilesSurface } from "./files-surface.ts";
 import { createSemanticThemeSnapshot } from "./theme.ts";
 import { terminalDisplayWidth } from "./panel-host.ts";
