+++ packages/daemon/src/tui/mirror/home-surface.tsx
@@ -3,7 +3,7 @@ import { For, Show } from "solid-js";
 import { rollupChips, type FleetRollup } from "../team/home.ts";
 import { STATUS_GLYPH } from "./status-grammar.ts";
 import type { SemanticThemeSnapshot } from "./theme.ts";
-import { ActionChip, EmptyState, SectionHeader, SelectableRow } from "./recipes.tsx";
+import { ActionChip, EmptyState, SectionHeader, SelectableRow } from "./recipe-components.tsx";
 import {
   homeActionAtProjection,
   type HomeActionId,
