+++ packages/daemon/src/tui/mirror/runtime/production-data-path.test.ts
@@ -60,11 +60,11 @@ describe("production OpenTUI data path", () => {
         "packages/daemon/src/tui/mirror/features/missions-activity/feature.tsx",
         "packages/daemon/src/tui/mirror/features/dialogs/feature.tsx",
         "packages/daemon/src/tui/mirror/features/settings/feature.ts",
-        "packages/daemon/src/tui/mirror/files-surface.tsx",
-        "packages/daemon/src/tui/mirror/changes-surface.tsx",
+        "packages/daemon/src/tui/mirror/files-surface-view.tsx",
+        "packages/daemon/src/tui/mirror/changes-surface-view.tsx",
         "packages/daemon/src/tui/mirror/missions-surface.tsx",
-        "packages/daemon/src/tui/mirror/activity-surface.tsx",
-        "packages/daemon/src/tui/mirror/workspace/command-palette-surface.tsx",
+        "packages/daemon/src/tui/mirror/activity-surface-view.tsx",
+        "packages/daemon/src/tui/mirror/workspace/command-palette-surface-view.tsx",
         "packages/daemon/src/tui/mirror/widget-surface.tsx",
       ]),
     );
