+++ packages/daemon/src/tui/mirror/workspace/command-palette-surface-view.tsx
